backend/pyproject.toml
[project]
name = "reviewcopies"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django-filter>=24.3",
    "django-rest-knox>=5.0.2",
    "django>=5.1,<5.2",
    "djangorestframework-simplejwt>=5.3.1",
    "djangorestframework>=3.15,<3.16",
    "faker>=30.8.2",
]

[tool.uv]
dev-dependencies = [
    "django-extensions>=3.2.3",
    "ipython>=8.29.0",
    "pre-commit>=4.0.1",
    "pyyaml>=6.0.2",
]
